fix(cdk/drag-drop): move type DragConstrainPosition in DragRef
fix circular dependency
